auto git add all

+++ alexnet_training_simple.py
@@ -260,7 +260,7 @@ def main(argv = None):
 
             accuracy = tf.reduce_mean(tf.cast(correct_prediction, tf.float32))
 
-            top_5_indices = tf.nn.top_k(score, k=5).indices
+            top_5_indices = tf.cast(tf.nn.top_k(score, k=5).indices, tf.float32)
             correct_prediction_top5 = tf.equal(top_5_indices, y)
             raw_acc = tf.cast(tf.reduce_sum(correct_prediction_top5, axis = 1), tf.float32)
             top_5_accuracy = tf.reduce_mean(raw_acc)
